app/requirements.txt
absl-py==1.0.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
flatbuffers==1.12
idna==3.3
kafka-python==2.0.2
numpy==1.22.3
opencv-contrib-python-headless==4.5.5.64
pandas==1.4.1
Pillow==9.0.1
pybind11==2.9.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
tflite-runtime==2.5.0
tflite-support==0.3.1
urllib3==1.26.9
